Unity example improvements
