+++ examples/csharp/HelloworldUnity/Assets/Scripts/HelloWorldTest.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8c088e5dee11c45fc95e41b9281d55e2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
